refactor: :recycle: move image provider logic into composable
